# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Микросервисная система управления заказами
  description: API для управления пользователями и заказами
  version: 1.0.0
servers:
  - url: http://localhost:8000/v1
    description: Локальный сервер разработки
  - url: http://api_gateway:8000/v1
    description: Docker окружение
  - url: http://production.example.com/v1
    description: Production окружение
...
paths:
  /users/register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      description: Создание нового пользователя в системе
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  example: password123
                name:
                  type: string
                  example: Иван Иванов
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  email: user@example.com
                  name: Иван Иванов
                  roles: ["user"]
                  createdAt: "2024-01-01T12:00:00.000Z"
                  updatedAt: "2024-01-01T12:00:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/login:
    post:
      tags:
        - Authentication
      summary: Вход в систему
      description: Аутентификация пользователя и получение JWT токена
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                data:
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    email: user@example.com
                    name: Иван Иванов
                    roles: ["user"]
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users:
    get:
      tags:
        - Users
      summary: Список пользователей
      description: Получение списка пользователей (только для администраторов)
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: email
          in: query
          schema:
            type: string
          description: Фильтр по email
        - name: role
          in: query
          schema:
            type: string
          description: Фильтр по роли
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /orders:
    post:
      tags:
        - Orders
      summary: Создать заказ
      description: Создание нового заказа
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/OrderItem'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags:
        - Orders
      summary: Список заказов
      description: Получение списка заказов текущего пользователя
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [created, in_work, completed, cancelled]
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
          description: Фильтр по пользователю (только для администраторов)
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, total]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid email format
...
    OrderItem:
      type: object
      required:
        - product
        - quantity
        - price
      properties:
        product:
          type: string
          example: Товар 1
        quantity:
          type: integer
          minimum: 1
          example: 2
        price:
          type: number
          minimum: 0
          example: 100.50
...
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/Error'
    Forbidden:
      description: Доступ запрещен
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/Error'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/Error'
    Conflict:
      description: Конфликт данных
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                $ref: '#/components/schemas/Error'